Crypto Prices Monitor/requirements.txt
kafka-python
psycopg2
schedule
requests
docker
pytest
sqlalchemy
pandas
python-dotenv